msmtp_enqueue: each time it's called it enqueue mails in ~/.msmtp-queue/
for each message three files are created with the same base (a timestamp).
* foo.args: command line arguments used
* foo.msg: stdin data, the actual message
* foo.msg.pid: pid of who's writting foo.msg, and deleted after it finishes properly

Only mails without the corresponding `.msg.pid` are supposed to be sent. Also, if the
`.msg.pid` still there we can use it's content, and check /proc/n/fd/ to see if it's
a zombie message which needs to be trashed.